ml_service/ml_service/services/weather.py
import requests

import os
from dotenv import load_dotenv

load_dotenv()

API_KEY = os.getenv("OPENWEATHER_API_KEY")
def get_weather(lat, lon):
    url = f"https://api.openweathermap.org/data/2.5/weather?lat={lat}&lon={lon}&appid={API_KEY}&units=metric"
    res = requests.get(url).json()

    weather = res["weather"][0]["main"]
    temp = res["main"]["temp"]

    return weather, temp