ml_service/ml_service/requirements.txt
flask
pandas
numpy
scikit-learn
joblib
requests
python-dotenv